data size converted to long

+++ gpu_manager.cpp
@@ -8,7 +8,7 @@
 #include <iostream>
 
 /*Get the number of GPU devices present on the host */
-gpu_manager::gpu_manager(int batch_size, int dimension){
+gpu_manager::gpu_manager(long batch_size, int dimension){
     this->batch_size = batch_size;
     this->dimension = dimension;
 
@@ -31,7 +31,7 @@ gpu_manager::gpu_manager(int batch_size, int dimension){
 void gpu_manager::copy_data(double* host_Mat, double* host_Vect) {
     std::cout << "========================Copying data for all GPUs======================================" << std::endl;
     for(int i=0; i<list_of_users.size(); i++){
-        int offset = compute_batch_offset(i);
+        long offset = compute_batch_offset(i);
         list_of_users.at(i)->set_device(i,"Data Copy");
         list_of_users.at(i)->copy_to_device(host_Mat + offset*dimension, host_Vect);
     }
@@ -44,14 +44,14 @@ float gpu_manager::compute_and_store(double* host_ResVect) {
         list_of_users.at(i)->set_device(i, "Compute and Store");
         list_of_users.at(i)->start_event();
         list_of_users.at(i)->launch_kernel();
-        int offset = compute_batch_offset(i);
+        long offset = compute_batch_offset(i);
         list_of_users.at(i)->copy_to_host(host_ResVect + offset);
         time_sec = std::max(time_sec, list_of_users.at(i)->stop_event());
     }
     return time_sec;
 }
 
-int gpu_manager::compute_batch_offset(int batch_id){
+long gpu_manager::compute_batch_offset(int batch_id){
     return batch_id * batch_matRowSize;
 }
